Implement custom .dys format and colorize Board#toString

+++ src/main/scala/net/akouryy/kreins/strategy/DyagsekiSearch.scala
@@ -5,7 +5,7 @@ import game.{Board, Pos}
 
 import scala.util.Random
 
-class DyagsekiSearch(val pt: Map[Board, List[(Byte, Byte)]]) {
+class DyagsekiSearch(val pt: Map[Board, List[(Byte, Int)]]) {
   def bestMove(board: Board): Option[Int] = {
     val pp = board.possPlaceable.code
 
@@ -21,7 +21,7 @@ class DyagsekiSearch(val pt: Map[Board, List[(Byte, Byte)]]) {
         )
       ) {
         for(ls <- pt.get(b)) {
-          var rand = Random.nextInt(ls.map(_._2.toInt).sum)
+          var rand = Random.nextInt(ls.map(_._2).sum)
           for((p0, _) <- ls.find { case (_, n) =>
             rand -= n
             rand <= 0
